Move most of lightmap to bevy_render
